Implement initialization of new persistent digest values
This enables migration to persistent value integrity checks via an OTA
update. When the new vbmeta is verified, if a persistent digest value
does not exist it is initialized with the current correct value. The
result is that a partition previously not integrity checked is now
verified against the contents at the point the feature is enabled. This
currently only works for partitions with hash descriptors.

Bug: 73020477
Test: unit
Change-Id: Iae709d31902cbcfe9f1e7e3f3d4e3b6dc783b716