libavb: Only query partition GUIDs when the cmdline needs them.
On devices using BOARD_AVB_VBMETA_MAINLINE, the kernel command line
should not contain the $(ANDROID_SYSTEM_PARTUUID) pattern, since the
device will be using an initial ramdisk for both normal boot and
recovery. Such a device will also not have system partition entries
in the GPT, and thus querying the GUID will fail.

In this case we want avb_sub_cmdline() to only query the partition GUID
if the search string is actually present.

Bug: 116010959
Test: New unit test and all unit tests pass
Change-Id: I0dded7026423df0481076239e2e3ee4feda29af5