avbtool: Use no salt instead of random salt for persistent digest descriptors
Descriptors that use a persistent digest (where the digest value is
stored on device in persistent storage and the descriptor has an empty
digest field) cannot use a random salt generated by avbtool because that
would change the digest value. This switches the default (if no '--salt'
option is provided) to no salt if --use_persistent_digest is used with
add_hash(tree)_footer.

Fixes: 118402832
Test: atest libavb_host_unittest
Change-Id: I5115670fe013ffcc013a3b542f70a22f3d7e380e