avbtool: verify_image: Use openssl command instead of PyCrypto.
The build servers are now running 'avbtool verify_image' for
validation purposes (yay!) but currently this requires PyCrypto to be
installed. We can avoid this dependency by shelling out to the
openssl(1) command, just like we do for signing.

Existing unit tests (AvbToolTest.VerifyImage*) covers both positive
and negative cases so this change is pretty safe.

Bug: 119624011
Test: Unit tests pass.
Change-Id: Ie34f4c7609423ec6b8937dcc75c5f1de66dd5f3d