at_auth_unlock.py: Fix exception when using Python2 w/ non-unlock cred in directory
zipfile.BadZipFile only exists on Python 3. On Python 2 it's
'zipfile.BadZipfile'. This leads to an AttributeError exception if the
user provides a directory argument that contains a file that's not a
valid unlock credential, which is valid usage.

Unit tests didn't catch this because they only run with Python 3, due to
the usage of unittest.mock which doesn't exist in Python 2.7.

Fixes: 120501195
Test: Verified scenario above no longer results in exception on Python
2.7.13

Change-Id: I2d3681f99dccf5aaa4f5b7512b729eec5cd636cf