avbtool: Add 'zero_hashtree' command.
This is useful for trading compressed image size for having to
reculculate the hashtree and FEC at runtime. If this is done all of
the hashtree and FEC data is set to zero except for the first eight
bytes which is set to the magic `ZeRoHaSH`. Applications can use this
to detect if recalculation is needed.

Also add a --accept_zeroed_hashtree option to 'avbtool verify_image'
which if given will treat a zeroed hashtree as equivalent to a valid
hashtree.

Bug: 130317158
Test: New unit tests + all tests pass.
Change-Id: I00054145942f1ffacfa77b0e96c787bf0c1a9dba