Remove algorithm field from AvbIcpHeader
This field is currently not used. It may also bring confusion as
multiple logs may use different algorithms. If required, it can later
be reinstated in the AvbIcpEntry class instead of AvbIcpHeader.

Remove part of the unit test that was comparing byte-to-byte the
serialization. This is brittle as any code change would require to
update the test.

Test: atest --host avbtool_test
Test: atest --host libavb_host_unittest
Change-Id: I5190443d761449bc574cb3d1de88cc571211e919