Fix lookup of AftlDescriptor before validation
AftlDescriptor is located after the AvbVBMetaImageHeader structure on
the vbmeta partition. This is not a standard descriptor and therefore,
it is not contained in the AvbSlotVerifyData vbmeta_images.

Change the logic to find the descriptor by reading the partition where
the VBMeta was found. Update the unit tests to handle the new function
definition. Fix the order of inclusion and location of headers.

Test: atest --host libavb_host_unittest
Bug: 150587998, 150586478
Change-Id: I3bc9d716e96d1afca6b65cc5f71ebbd5ec39d21b