Porting aftltool and its test suites to Python 3.
As of January 1st, 2020 Python 2 is not officially supported anymore:
https://www.python.org/dev/peps/pep-0373/#id4

As a consequence aftltool has been ported to Python 3. This change
breaks backward compatibility with Python 2 runtime environments.

Cleaned up use of str() and bytes(). Standardize on using bytes()
instead of bytearray() except for situations where the mutable
property of bytesarray() is actually used. Also fixed lint errors.

One issue still open is that Python 3 does not support relative import
of modules anymore which causes importing of generated protobuf _*pb2
modules to fail. At the moment there is a workaround in place until
a solution has been found as part of b/154068467.

Bug: 153057071
Test: atest --host libavb_host_unittest
Test: atest --host aftltool_test
Test: ./aftltool_integration_test.py
Change-Id: Ia4053d2cd81d579004c6fda90bc32616b9674f88