Fix preload partition failure behavior
This fixes a bug when handling an inconsistency between the image size
reported by the vbmeta and the preloaded size.

The docs for the `get_preloaded_partition()` callback specify that
`out_num_bytes_preloaded` must be no larger than `num_bytes`.

If the preloaded partition is smaller than the num_bytes requested, the
correct behavior seems to be to return what was preloaded to prevent
libavb from reading the partition again. With this callback behavior,
avb_slot_verify() segfaults after trying to free the buffer returned by
get_preloaded_partition().

Test: Added test which verifies graceful failure instead of segfault.

Bug: fxb/40064
Change-Id: I798cedba2f6f1ccb19f25735202894efbabed09f